refactor(rpc): split jsonrpc handlers by namespace (backlog #11 phase 2)
The 900-line match in jsonrpc_handler is now a prefix-dispatch to
four namespace modules:

  * eth.rs       — 20 methods (chainId, blockNumber, gasPrice,
                   estimateGas, getBalance, getTransactionCount,
                   getBlockByNumber, getBlockByHash,
                   getTransactionByHash, getTransactionReceipt,
                   getBlockReceipts, sendRawTransaction, call,
                   getLogs, feeHistory, maxPriorityFeePerGas,
                   syncing, accounts, getCode, getStorageAt)
  * net.rs       — 2 methods (version, listening)
  * web3.rs      — 1 method (clientVersion)
  * sentrix.rs   — 7 methods (sendTransaction, getBalance,
                   getValidatorSet, getDelegations,
                   getStakingRewards, getBftStatus,
                   getFinalizedHeight)

Each module exposes

    pub(super) async fn dispatch(
        method: &str,
        params: &Value,
        state: &SharedState,
    ) -> DispatchResult

where DispatchResult = Result<Value, (i32, String)> is now a
crate-level type. mod.rs routes by method prefix and wraps the
result into JsonRpcResponse once.

Error paths were converted from the old
'return Json(JsonRpcResponse::err(id, code, msg))' pattern to
'Err((code, msg))'. The envelope wrap happens once in the
top-level handler — cleaner and removes the need to pass 'id'
into every error site.

No behaviour change. All workspace tests pass, clippy clean,
existing integration tests (including the jsonrpc batch-size
guard, sentrix native RPC coverage, and eth_getBlockReceipts
coverage) continue to exercise the same code paths through the
new dispatch layer.
